refactor(endpoint-micropub): split jf2 and mf2 functions
