+++ packages/endpoint-micropub/lib/controllers/query.js
@@ -1,6 +1,6 @@
 import Debug from 'debug';
 import httpError from 'http-errors';
-import {getMf2Properties, jf2ToMf2, url2Mf2} from '../microformats.js';
+import {getMf2Properties, jf2ToMf2, url2Mf2} from '../mf2.js';
 import {getConfig, queryList} from '../query.js';
 
 const debug = new Debug('indiekit:error');
